move token exchange function into shared mondoo utils package
